refactor: move WithFilters para commons
