+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/ListContainer.tsx
@@ -1,5 +1,6 @@
 import { WithPagination } from "@/features/acf/common/frontend/WithPagination";
 import { WithSearch } from "@/features/acf/common/frontend/WithSearch";
+import { WithFilters } from "@features/acf/common/frontend/WithFilters";
 import { WithSorting } from "@features/acf/common/frontend/WithSorting";
 import React from "react";
 import type { AcfDashboardType } from "../../../../../../common/model";
@@ -9,7 +10,6 @@ import { CoeqDataTable } from "./modules/CoeqDataTable";
 import { FiltersBarCoeq } from "./modules/FiltersBar";
 import { InternalCardsCoeq } from "./modules/Internalcards";
 import { ToolBar } from "./modules/ToolBar";
-import { WithFilters } from "./modules/WithFilters";
 
 export type ListContainerProps = {
     list: AcfDashboardType;
